+++ include/settings.h
@@ -1,0 +1,30 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   settings.h                                         :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/05/28 08:45:22 by freddy            #+#    #+#             */
+/*   Updated: 2024/05/28 09:47:08 by freddy           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#ifndef SETTINGS_H
+# define SETTINGS_H
+
+#include <stdbool.h>
+
+// General
+
+# define GREETING true
+# define VERBOSE true
+# define DEBUG true
+
+// Window Setup
+
+# define START_HEIGHT 540
+# define START_WIDTH 960
+# define WINDOW_NAME "cub3d"
+
+#endif