+++ src/cub3d.c
@@ -6,7 +6,7 @@
 /*   By: freddy <freddy@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/21 07:33:54 by fschuber          #+#    #+#             */
-/*   Updated: 2024/05/24 13:42:33 by freddy           ###   ########.fr       */
+/*   Updated: 2024/05/28 10:18:58 by freddy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,5 +35,8 @@ int	main(int argc, char **argv)
 		generate_map();
 		parse_input("generated_maze.cub");
 	}
+	setup_mlx();
+	mlx_loop(get_persistent_data()->mlx);
+	mlx_terminate(get_persistent_data()->mlx);
 	return (EXIT_SUCCESS);
 }